# Optimize String#casecmp? for ASCII strings