@casperisfine 2020-07-29
@XrXr @tenderlove would you have time to do a preliminary review of this? Like any obvious things I might have missed?