@casperisfine 2020-07-30
> Couldn’t you shrink the diff for review?

Not sure what you mean.