@nurse 2021-01-24
This breaks spec when the encoding is legacy single byte encodings like ISO-8859-1. I added more test at https://github.com/ruby/ruby/commit/1bd27a78827109f85357f80827a9975fae9bb2a5.

When you avoid that case, you have a option around `coderange`: `coderange` is a cached information whether the string contains (1) only ASCII 7 bit characters (2) also has 8 bit characters (3) broken byte sequence (4) unknown. Some strings are already scanned its coderange and caches it in a string object, but others are not. Whether this `casecmp?` optimization uses the cache and not scan string if the cache doesn't exist, or scan if it doesn't have a cache. If you use the cache, I wonder whether strings in real applications have cache or not. If you scan, I wonder if it still gets faster.