@casperisfine 2021-01-25
Thanks for the extra test case, I'll see if I can handle it.