@casperisfine 2021-01-25
Ok, so I pushed another commit to restrict the optimization to strings with 7BIT coderange. The issue was that `TOLOWER` only works on `A..Z`:

```c
static inline int
rb_isupper(int c)
{
    return 'A' <= c && c <= 'Z';
}

static inline int
rb_tolower(int c)
{
    return rb_isupper(c) ? (c|0x20) : c;
}
```

Even though it could work for you specific test:

```ruby
>> 0xC0 | 0x20 == 0xE0
=> true
```

But I suppose it can't be encoding aware.

> I wonder whether strings in real applications have cache or not.

Hard to tell, but since `casecmp?` returns a boolean, you'd think very often one of the two strings is a constant, which kinda limits the impact of scanning for coderange.

> If you scan, I wonder if it still gets faster.

We could probably benchmark that. I think I remember some benchmark that did clear the coderange, I don't remember how though.