@nobu 2021-01-25
Please insert a blank line before the last table in the commit log.
RDoc's markdown parser doesn't consider it a code block.