@casperisfine 2021-01-25
> Please insert a blank line before the last table in the commit log.

Done.

I also updated the description a bit, since the `strncmp` part is outdated.