@Earlopain 2026-08-16
Probably not needed anymore https://github.com/ruby/ruby/pull/17444